[Bleeding] Store correct block type when hitting blocks. Fixes BUKKIT-5209
The block obtained earlier within the tick does not reflect the block
at the current location, rather it is the block at the location from the
previous tick. Thusly when the variable is used to check if the arrow
is still in the block, it fails and "hits" the block a second time.