Always deduct XP when enchanting an item. Fixes BUKKIT-5557
The previous behavior was to only deduct XP if enchanting a non-book,
or only when en enchantment as actually added to a book.

The behavior after commit ac1a2d0233 was that XP was not deducted for
items or books if an enchantment was not added.

I have chosen to make the consistent behavior that XP is always
deducted, though if this is not desirable, it would be preferable to
restore the old behavior for non-book items, at least.