Store arbitrary extra data in CraftMetaItem.
Only data that is not already well-known ("auto-registered" by an existing ItemMetaKey definition) will be stored.

If no such data is present, as will be the normal case, each CraftMetaItem will have one extra empty reference (to a Map, which is not initialized until it needs to hold data).

Any extra metadata is serialized in a backwards-compatible way, and properly copied to other items, and applied to and from the underlying NBT data.

This is not meant to be an item data storage API, though it could be exposed in Bukkit as such. I would suggest using a Configuration interface wrapper for getting and setting extra data, if API access to extra item data was allowed and desired.

The primary purpose of this change is to avoid losing data that Bukkit does not know about. This could be due to Mojang adding new item types or fields that Bukkit has not yet created a mirror class definition for, or from plugins that add their own custom data (which, yes, I admit would have to be done outside of the API, at present).

Plugin issues aside, this could be a valuable change moving forward as Mojang adds more item classes with custom data (perhaps), or starts generalizing the structure in the move towards custom item and block types. Making sure data is at least properly stored and transferred could prevent a lot of potential migration issues while dev builds are updating.

As for plugins, this change clears up a number of issues with plugins that use custom data, including interoperability with other plugins, creative mode and various inventory events.